requirements.txt
Flask==1.1.2
gunicorn==19.10.0
google-cloud-datastore==1.15.3